src/main/java/org/example/widgetsapi/entity/Widget.java
package org.example.widgetsapi.entity;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Getter;

import java.time.LocalDateTime;
import java.util.Objects;
import java.util.Optional;
import java.util.UUID;

@Getter
@AllArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Widget implements Comparable<Widget> {

    private final UUID id;
    private final Point coordinates;
    private final Optional<Integer> zindex;
    private final int width, height;
    private final LocalDateTime lastModificationDate;

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Widget widget = (Widget) o;
        return id.equals(widget.id);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }

    @Override
    public int compareTo(Widget o) {
        return zindex.get() - o.zindex.get();
    }
}


